Rewrote bencode to be a duplex stream
`require('bencode')` now exports a duplex stream
consisting of an encoder and decoder stream.

While the handling is still quite flaky, it supports
piping to *AND* from it in different manners.
It looks at if the source/destination is an object stream
and decides where to/from to pipe it. Caution is advised.
If in doubt, or not working use decoder and encoder streams separately.